% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_data_get.R
\name{js_data_get.js_data}
\alias{js_data_get.js_data}
\title{js_data_get}
\usage{
js_data_get.js_data(obj, order = 1)
}
\arguments{
\item{obj}{a js_data object from get_obj funtion}

\item{order}{the order of the object in the database from left to right apart from object ID, default is 1}
}
\description{
This function is to transfer the json object retrieved from get_obj function to a dataframe.
}
