% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js_data_ret.R
\name{js_data_ret}
\alias{js_data_ret}
\title{Original function}
\usage{
js_data_ret(obj, order = 1)
}
\arguments{
\item{obj}{a js_data object from get_obj funtion}

\item{order}{the order of the object in the database from left to right apart from object ID, default is 1}
}
\description{
Original function
}
