% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obj.R
\name{get_obj}
\alias{get_obj}
\title{get_obj}
\usage{
get_obj(web, query, ID, KEY, returnjs = T)
}
\arguments{
\item{web}{Character. Your leancloud database RESTFUL API. You can find it in leancloud website-console-setting-app keys}

\item{query}{Character. The name of the object that you want to retrieve}

\item{ID}{AppID. Character. You can find it in leancloud website-console-setting-app keys}

\item{KEY}{AppKey. Character. You can find it in leancloud website-console-setting-app keys}

\item{returnjs}{Boolean.This paradigm indicate whether return a js_data object for further data extraction}
}
\description{
This function is a simple wrapper of leancloud RESTful API to retrieve a value or a object from leancloud database.
}
